Importing CodePen CSS
When putting the CSS that was originially here on the CodePen version of the page, the same issue (inactive streamers' elements spilling out of their division on one side) was reproduced. However, the original CSS on the CodePen version did not have that issue, so I copy and pasted it here to see if maybe something in that code fixes/prevents the issue.